+++ phidgets_magnetometer/include/phidgets_magnetometer/magnetometer_ros_i.hpp
@@ -59,7 +59,7 @@ class MagnetometerRosI final : public rclcpp::Node
         magnetometer_pub_;
     void timerCallback();
     rclcpp::TimerBase::SharedPtr timer_;
-    int publish_rate_;
+    double publish_rate_;
 
     rclcpp::Time ros_time_zero_;
     bool synchronize_timestamps_{true};
